+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Head blocks textures are now loading instantly and do not cause server to freeze at the start now
 * Fixed /bb top was reversed while MySQL was enabled
 * Disabled game end rewards by default because they were confusing for users
+* Fixed server was stopped using Shutdown-When-Game-Ends option but player just left not started arena
 
 ### 3.5.1 Release (17.08.2018)
 * Fixed NumberFormatException for language.yml migrator - this error is very rare to occur in normal environment but it was reported so I fixed it