+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## BuildBattle 3 changelog
 
+### 4.0.0 Beta build 3 (03.10.2018)
+* Removed annoying "Please enable bStats" message
+
 ### 4.0.0 Release (20.08.2018 - 28.09.2018)
 * Fixed NullPointerException for users who where no longer online
 * Now using default values when they not exist in config to avoid NullPointerException from ConfigPreferences